mejoras en clean fish

+++ clean_fish_history.rb
@@ -3,6 +3,10 @@
 # Especifica la ruta del archivo
 ruta = "~/.local/share/fish/fish_history"
 
-# abre el archivo y elimina el contenido
-File.open(File.expand_path(ruta), "w") {}
-puts "clean history"
+if File.exist?(ruta)
+  # abre el archivo y elimina el contenido
+  File.open(File.expand_path(ruta), "w") {}
+  puts "clean history"
+else
+  puts "el archivo no existe"
+end
